+++ routes/api.php
@@ -5,7 +5,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/user', function (Request $request) {
-    return $request->user();
+    return $request->user()->load('subscription.plan');
 })->middleware('auth:sanctum');
 
 // User Billing & Plans